o1o

+++ week02/F.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-long int recursive_factorial(int n){
+long long recursive_factorial(int n){
     if (n == 0) return 1;
     else return n * recursive_factorial(n-1);
 }
